+++ offlineeuro/src/main/java/nl/tudelft/trustchain/offlineeuro/community/OfflineEuroCommunity.kt
@@ -38,6 +38,7 @@ import nl.tudelft.trustchain.offlineeuro.cryptography.RandomizationElementsBytes
 import nl.tudelft.trustchain.offlineeuro.entity.TransactionDetailsBytes
 import nl.tudelft.trustchain.offlineeuro.enums.Role
 import java.math.BigInteger
+import android.util.Log
 
 object MessageID {
     const val GET_GROUP_DESCRIPTION_CRS = 9
@@ -252,7 +253,7 @@ class OfflineEuroCommunity(
         challenge: BigInteger,
         publicKeyBytes: ByteArray,
         bankPublicKeyBytes: ByteArray,
-        amount: Double
+        amount: Long
 
     ) {
         val bankPeer = getPeerByPublicKeyBytes(bankPublicKeyBytes)
@@ -390,6 +391,7 @@ class OfflineEuroCommunity(
     }
 
     fun onTransactionPacket(packet: Packet) {
+        //Log.d("OfflineEuroCommunity", "On transaction packet!")
         val (peer, payload) = packet.getAuthPayload(TransactionDetailsPayload)
         onTransaction(peer, payload)
     }
@@ -398,6 +400,7 @@ class OfflineEuroCommunity(
         peer: Peer,
         payload: TransactionDetailsPayload
     ) {
+        //Log.d("OfflineEuroCommunity", "On transaction message!")
         val publicKey = payload.publicKey
         val transactionDetailsBytes = payload.transactionDetailsBytes
         val message = TransactionMessage(publicKey, transactionDetailsBytes, peer)
